requirements.txt
streamlit==1.31.0
sentence-transformers==2.3.1
faiss-cpu
pandas
numpy
